+++ handlers/producer/tx-sender/producer.go
@@ -18,7 +18,8 @@ func PrepareMsg(txctx *engine.TxContext, msg *sarama.ProducerMessage) error {
 	b, ok := txctx.Get("invalid.nonce").(bool)
 	switch {
 	case len(txctx.Envelope.GetErrors()) == 0 && ok && b:
-		msg.Topic = viper.GetString(broker.TxNonceViperKey)
+		// If nonce is invalid we redirect envelope to tx-crafter
+		msg.Topic = viper.GetString(broker.TxCrafterViperKey)
 		p = txctx.Envelope.TxEnvelopeAsRequest()
 	case !txctx.Envelope.OnlyWarnings():
 		msg.Topic = viper.GetString(broker.TxRecoverViperKey)