+++ CHANGELOG.md
@@ -40,6 +40,7 @@ All notable changes to this project will be documented in this file.
  
 ### ⚠ BREAKING CHANGES
  * Remove the `tx-decoder` microservice and decode logs in the tx-listener. The tx-listener publish transactions directly in the `topic-tx-decoded`
+ * Merge the `tx-nonce` microservice into `tx-crafter` microservice. Consequently topic `tx-nonce` has been deprecated
  * Instead of producing and consuming envelopes to Orchestrate, a user will produce `TxRequest` and only consume `TxResponse`
  * Remove `/v1` prefix in the http rest path for the envelope-store and chain-registry
  * The `tx-listener` produces kafka messages only in the topic `topic-tx-decoded` instead of the `topic-tx-decoder-{chainID}`